+++ src/components/FeaturePanel/TagsTable/Wrapper.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  position: relative;
+  margin-bottom: 2em;
+`;
